move shareRecipe function
